SAP Doc Monitor/sap-doc-monitor/requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
webdriver-manager>=3.8.0
APScheduler>=3.10.0
Flask>=2.3.0
google-cloud-storage>=2.10.0
